+++ src/js/methods.js
@@ -256,9 +256,9 @@ export default {
       const newHeight = naturalHeight * ratio;
 
       if (dispatchEvent(this.element, EVENT_ZOOM, {
-        originalEvent: _originalEvent,
+        ratio,
         oldRatio: width / naturalWidth,
-        ratio: newWidth / naturalWidth,
+        originalEvent: _originalEvent,
       }) === false) {
         return this;
       }
@@ -403,9 +403,20 @@ export default {
       const ratio = imageData.width / imageData.naturalWidth;
 
       forEach(data, (n, i) => {
-        n /= ratio;
-        data[i] = rounded ? Math.round(n) : n;
+        data[i] = n / ratio;
       });
+
+      if (rounded) {
+        // In case rounding off leads to extra 1px in right or bottom border
+        // we should round the top-left corner and the dimension (#343).
+        const bottom = Math.round(data.y + data.height);
+        const right = Math.round(data.x + data.width);
+
+        data.x = Math.round(data.x);
+        data.y = Math.round(data.y);
+        data.width = right - data.x;
+        data.height = bottom - data.y;
+      }
     } else {
       data = {
         x: 0,
@@ -743,8 +754,6 @@ export default {
       dstHeight = srcHeight;
     }
 
-    // All the numerical parameters should be integer for `drawImage`
-    // https://github.com/fengyuanchen/cropper/issues/476
     const params = [
       srcX,
       srcY,
@@ -764,6 +773,8 @@ export default {
       );
     }
 
+    // All the numerical parameters should be integer for `drawImage`
+    // https://github.com/fengyuanchen/cropper/issues/476
     context.drawImage(source, ...params.map(param => Math.floor(normalizeDecimalNumber(param))));
 
     return canvas;